feat: seeds home page, clickable card (#4737)
