+++ shared-helpers/index.ts
@@ -26,6 +26,7 @@ export * from "./src/utilities/stringFormatting"
 export * from "./src/utilities/token"
 export * from "./src/utilities/unitTypes"
 export * from "./src/views/components/BloomCard"
+export * from "./src/views/components/ClickableCard"
 export * from "./src/views/CustomIconMap"
 export * from "./src/views/forgot-password/FormForgotPassword"
 export * from "./src/views/layout/ExygyFooter"
