+++ shared-helpers/src/locales/ar.json
@@ -1147,6 +1147,7 @@
   "welcome.seeRentalListings": "انظر جميع الايجارات",
   "welcome.signUp": "الحصول على تنبيهات كلما تم نشر قائمة جديدة",
   "welcome.signUpToday": "سجّل اليوم",
+  "welcome.subtitle": "موردك للعثور على مساكن متعددة العائلات بناءً على دخلك واحتياجات أسرتك",
   "welcome.title": "تقدم بطلب للحصول على سكن ميسور التكلفة في",
   "welcome.viewAdditionalHousing": "عرض فرص وموارد الإسكان الإضافية",
   "whatToExpect.default": "سيتم الاتصال بالمتقدمين من قبل وكيل الملكية بترتيب الترتيب حتى يتم ملء الشواغر. سيتم التحقق من جميع المعلومات التي قدمتها وتأكيد أهليتك. ستتم إزالة طلبك من قائمة الانتظار إذا قدمت أي بيانات احتيالية. إذا لم نتمكن من التحقق من تفضيل السكن الذي طالبت به ، فلن تحصل على الأفضلية ولكن لن يتم معاقبتك بطريقة أخرى. إذا تم اختيار طلبك ، فكن مستعدًا لملء طلب أكثر تفصيلاً وتقديم المستندات الداعمة المطلوبة.",
