+++ shared-helpers/src/locales/bn.json
@@ -1073,6 +1073,7 @@
   "welcome.seeRentalListings": "সব ভাড়া দেখুন",
   "welcome.signUp": "কোনো নতুন তালিকা পোস্ট করার সাথে-সাথে বার্তা পান",
   "welcome.signUpToday": "আজই সাইন-আপ করুন",
+  "welcome.subtitle": "আপনার আয় এবং পরিবারের চাহিদার উপর ভিত্তি করে বহু-পরিবারের আবাসন খুঁজে পেতে আপনার সম্পদ",
   "welcome.title": "সাশ্রয়ী মূল্যের আবাসনের জন্য আবেদন করুন",
   "welcome.viewAdditionalHousing": "অতিরিক্ত আবাসন সুযোগ এবং সম্পদ দেখুন",
   "whatToExpect.default": "শূন্যপদ পূরণ না হওয়া পর্যন্ত আবেদনকারীরা র agent্যাঙ্ক ক্রমে সম্পত্তি এজেন্টের সাথে যোগাযোগ করবে। আপনার দেওয়া সমস্ত তথ্য যাচাই করা হবে এবং আপনার যোগ্যতা নিশ্চিত করা হবে। যদি আপনি কোন প্রতারণামূলক বিবৃতি দিয়ে থাকেন তাহলে আপনার আবেদনটি প্রতীক্ষার তালিকা থেকে সরিয়ে দেওয়া হবে। যদি আমরা আপনার দাবি করা আবাসন পছন্দ যাচাই করতে না পারি, তাহলে আপনি অগ্রাধিকার পাবেন না কিন্তু অন্যথায় শাস্তি পাবেন না। আপনার আবেদন নির্বাচন করা উচিত, আরো বিস্তারিত আবেদন পূরণ এবং প্রয়োজনীয় সহায়ক নথি প্রদান করার জন্য প্রস্তুত থাকুন।",
