+++ shared-helpers/src/locales/es.json
@@ -1003,6 +1003,7 @@
   "welcome.seeRentalListings": "Ver viviendas en alquiler",
   "welcome.signUp": "Reciba correos electrónicos cuando se publique un anuncio nuevo",
   "welcome.signUpToday": "Regístrese hoy",
+  "welcome.subtitle": "Su recurso para encontrar viviendas multifamiliares según sus ingresos y necesidades del hogar",
   "welcome.title": "Haga su solicitud de vivienda de precio accesible en",
   "welcome.viewAdditionalHousing": "Vea Oportunidades y recursos adicionales de vivienda",
   "welcome.viewAdditionalHousingTruncated": "Ver oportunidades y recursos",
