+++ shared-helpers/src/locales/general.json
@@ -1001,6 +1001,7 @@
   "welcome.seeRentalListings": "See Rentals",
   "welcome.signUp": "Get emailed whenever a new listing is posted",
   "welcome.signUpToday": "Sign up today",
+  "welcome.subtitle": "Your resource to find multi-family housing based on your income and household needs",
   "welcome.title": "Apply for affordable housing in",
   "welcome.viewAdditionalHousing": "View additional housing opportunities and resources",
   "welcome.viewAdditionalHousingTruncated": "View opportunities and resources",
