+++ shared-helpers/src/locales/tl.json
@@ -991,6 +991,7 @@
   "welcome.seeRentalListings": "Tingnan ang mga Renta",
   "welcome.signUp": "Makatanggap ng email sa tuwing magpo-post ng bagong listahan",
   "welcome.signUpToday": "Mag-sign up ngayon",
+  "welcome.subtitle": "Ang iyong mapagkukunan upang makahanap ng multi-family housing batay sa iyong kita at mga pangangailangan sa sambahayan",
   "welcome.title": "Mag-apply para sa abot-kayang pabahay sa",
   "welcome.viewAdditionalHousing": "Tingnan ang karagdagang oportunidad at mapagkukunan ng pabahay",
   "welcome.viewAdditionalHousingTruncated": "Tingnan ang mga oportunidad at mapagkukunan",
