+++ shared-helpers/src/locales/vi.json
@@ -992,6 +992,7 @@
   "welcome.seeRentalListings": "Xem Thông tin Thuê nhà",
   "welcome.signUp": "Nhận email mỗi khi có danh sách mới được đăng",
   "welcome.signUpToday": "Đăng Ký Ngay Hôm Nay",
+  "welcome.subtitle": "Nguồn lực của bạn để tìm nhà ở nhiều gia đình dựa trên thu nhập và nhu cầu hộ gia đình của bạn",
   "welcome.title": "Ghi danh nhà ở giá phải chăng tại",
   "welcome.viewAdditionalHousing": "Xem Các Cơ hội Nhà ở và các Nguồn Hỗ trợ Bổ sung",
   "welcome.viewAdditionalHousingTruncated": "Xem các Cơ Hội và Nguồn Hỗ Trợ",
