+++ shared-helpers/src/locales/zh.json
@@ -992,6 +992,7 @@
   "welcome.seeRentalListings": "查看租賃部分",
   "welcome.signUp": "發佈新清單即收到電子郵件",
   "welcome.signUpToday": "立即註冊",
+  "welcome.subtitle": "根據您的收入和家庭需求尋找多戶住房的資源",
   "welcome.title": "申請此地的可負擔房屋：",
   "welcome.viewAdditionalHousing": "瀏覽其他住房機會和資源",
   "welcome.viewAdditionalHousingTruncated": "查看機會及資源",
