+++ shared-helpers/src/views/components/BloomCard.module.scss
@@ -22,7 +22,7 @@
 }
 
 .card-icon {
-  margin-bottom: var(--seeds-s3);
+  margin-block-end: var(--seeds-s3);
 }
 
 .card-heading-group {
