+++ sites/public/src/layouts/max-width.tsx
@@ -3,12 +3,19 @@ import styles from "./max-width.module.scss"
 
 interface MaxWidthLayoutProps {
   children: React.ReactNode
+  className?: string
 }
 
 const MaxWidthLayout = (props: MaxWidthLayoutProps) => {
   return (
-    <div className={styles["layout-max-width-container"]}>
-      <div className={styles["layout-max-width-content"]}>{props.children}</div>
+    <div className={styles["max-width-layout"]}>
+      <div
+        className={`${styles["layout-max-width-container"]} ${
+          props.className ? props.className : ""
+        }`}
+      >
+        <div className={styles["layout-max-width-content"]}>{props.children}</div>
+      </div>
     </div>
   )
 }
