+++ sites/public/src/patterns/CollapsibleSection.tsx
@@ -2,8 +2,8 @@ import React, { useState } from "react"
 import PlusIcon from "@heroicons/react/24/solid/PlusIcon"
 import MinusIcon from "@heroicons/react/24/solid/MinusIcon"
 import { Heading, Icon } from "@bloom-housing/ui-seeds"
-import styles from "./CollapsibleSection.module.scss"
 import { t } from "@bloom-housing/ui-components"
+import styles from "./CollapsibleSection.module.scss"
 
 interface CollapsibleSectionProps {
   /** All content under the title */
